+++ examples/base/src/pages/Demo/Detail/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import { ProDrawer, ProForm } from '@uiw-admin/components';
 import { Notify } from 'uiw';
 import { useSelector, useDispatch } from 'react-redux';
@@ -11,12 +11,10 @@ interface DetailProps {
   updateData?: any;
 }
 
-const Detail = (props: DetailProps) => {
-  const { updateData } = props;
+const Detail = ({ updateData }: DetailProps) => {
+  const baseRef = useRef<any>()
   const dispatch = useDispatch<Dispatch>();
-  const {
-    demo: { drawerVisible, tableType, queryInfo, isView },
-  } = useSelector((state: RootState) => state);
+  const { demo: { drawerVisible, tableType, queryInfo, isView } } = useSelector((state: RootState) => state);
 
   const onClose = () => dispatch({ type: 'demo/clean' });
 
@@ -37,8 +35,6 @@ const Detail = (props: DetailProps) => {
     },
   );
 
-  const dataSource: any = items(queryInfo, { isView });
-
   return (
     <ProDrawer
       width={800}
@@ -47,11 +43,24 @@ const Detail = (props: DetailProps) => {
       }
       visible={drawerVisible}
       onClose={onClose}
+      buttons={[
+        {
+          label: "保存",
+          type: "danger",
+          style: { width: 80 },
+          show: !isView,
+          onClick: () => baseRef?.current?.click()
+        }
+      ]}
     >
       <ProForm
         title="基础信息"
+        formType={isView ? 'pure' : 'card'}
+        submitRef={baseRef}
+        readOnly={isView}
+        readOnlyProps={{ column: 2 }}
         onSubmit={(initial, current) => {
-          const errorObj:any = {};
+          const errorObj: any = {};
           if (!current?.lastName) {
             errorObj.lastName = '名字不能为空';
           }
@@ -63,19 +72,12 @@ const Detail = (props: DetailProps) => {
           }
           mutate();
         }}
-        btnsContainer={{ justifyContent: "flex-start" }}
-        btns={[
-          {
-            btnType: 'submit',
-            label: '提交表单',
-            type: 'primary',
-          },
-        ]}
+        buttonsContainer={{ justifyContent: "flex-start" }}
         // 更新表单的值
         onChange={(initial: any, current: any) =>
           updateData({ queryInfo: { ...queryInfo, ...current } })
         }
-        formDatas={dataSource}
+        formDatas={items(queryInfo,isView) as any}
       />
     </ProDrawer>
   );